refactor: update for module refactoring
