+++ tests/ludwig/encoders/test_image_encoders.py
@@ -3,7 +3,7 @@ import torch
 
 from ludwig.encoders.image_encoders import MLPMixerEncoder, ResNetEncoder, Stacked2DCNN, ViTEncoder
 from ludwig.utils.misc_utils import set_random_seed
-from tests.integration_tests.utils import check_module_parameters_updated
+from tests.integration_tests.parameter_update_utils import check_module_parameters_updated
 
 RANDOM_SEED = 1919
