+++ tests/ludwig/encoders/test_sequence_encoders.py
@@ -14,7 +14,7 @@ from ludwig.encoders.sequence_encoders import (
     StackedTransformer,
 )
 from ludwig.utils.torch_utils import get_torch_device
-from tests.integration_tests.utils import check_module_parameters_updated
+from tests.integration_tests.parameter_update_utils import check_module_parameters_updated
 
 DEVICE = get_torch_device()
 RANDOM_SEED = 1919
